requirements.txt
notebook
PyGithub
python-dotenv
pandas
GitPython
matplotlib
seaborn
scipy
graphviz
PyYAML
toml